+++ sdk/monitor/azure-monitor-query/CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added additional `display_description` attribute to the `Metric` type.
 - Added a `MetricClass` enum to provide the class of a metric.
 - Added a `metric_class` attribute to the `MetricDefinition` type.
+- Added a `MetricNamespaceClassification` enum to support the `namespace_classification` attribute on `MetricNamespace` type.
 
 ### Breaking Changes